reorganise test submodules as per package submodules, create individual test files for each class
