+++ tests/test_creation/test_mathematical_combination.py
