+++ tests/test_outliers/test_winsorizer.py
@@ -1,12 +1,9 @@
-# Authors: Soledad Galli <solegalli@gprotonmail.com>
-# License: BSD 3 clause
-
-import pytest
 import numpy as np
 import pandas as pd
+import pytest
 from sklearn.exceptions import NotFittedError
 
-from feature_engine.outliers import Winsorizer, ArbitraryOutlierCapper, OutlierTrimmer
+from feature_engine.outliers import Winsorizer
 
 
 def test_Windsorizer(dataframe_normal_dist, dataframe_na, dataframe_vartypes):
@@ -163,151 +160,3 @@ def test_Windsorizer(dataframe_normal_dist, dataframe_na, dataframe_vartypes):
     with pytest.raises(NotFittedError):
         transformer = Winsorizer()
         transformer.transform(dataframe_vartypes)
-
-
-def test_ArbitraryOutlierCapper(dataframe_normal_dist, dataframe_na, dataframe_vartypes):
-    # test case 1: right end capping
-    transformer = ArbitraryOutlierCapper(max_capping_dict={'var': 0.10727677848029868}, min_capping_dict=None)
-    X = transformer.fit_transform(dataframe_normal_dist)
-
-    df_transf = dataframe_normal_dist.copy()
-    df_transf['var'] = np.where(df_transf['var'] > 0.10727677848029868, 0.10727677848029868, df_transf['var'])
-
-    # init params
-    assert transformer.max_capping_dict == {'var': 0.10727677848029868}
-    assert transformer.min_capping_dict is None
-    assert transformer.variables == ['var']
-    # fit params
-    assert transformer.right_tail_caps_ == {'var': 0.10727677848029868}
-    assert transformer.left_tail_caps_ == {}
-    assert transformer.input_shape_ == (100, 1)
-    # transform params
-    pd.testing.assert_frame_equal(X, df_transf)
-    assert X['var'].max() <= 0.10727677848029868
-    assert dataframe_normal_dist['var'].max() > 0.10727677848029868
-
-    # test case 2: both tails
-    transformer = ArbitraryOutlierCapper(max_capping_dict={'var': 0.20857275540714884},
-                                         min_capping_dict={'var': -0.19661115230025186})
-    X = transformer.fit_transform(dataframe_normal_dist)
-
-    df_transf = dataframe_normal_dist.copy()
-    df_transf['var'] = np.where(df_transf['var'] > 0.20857275540714884, 0.20857275540714884, df_transf['var'])
-    df_transf['var'] = np.where(df_transf['var'] < -0.19661115230025186, -0.19661115230025186, df_transf['var'])
-
-    # fit params
-    assert transformer.right_tail_caps_ == {'var': 0.20857275540714884}
-    assert transformer.left_tail_caps_ == {'var': -0.19661115230025186}
-    # transform params
-    pd.testing.assert_frame_equal(X, df_transf)
-    assert X['var'].max() <= 0.20857275540714884
-    assert X['var'].min() >= -0.19661115230025186
-    assert dataframe_normal_dist['var'].max() > 0.20857275540714884
-    assert dataframe_normal_dist['var'].min() < -0.19661115230025186
-
-    # test case 3: left tail
-    transformer = ArbitraryOutlierCapper(max_capping_dict=None, min_capping_dict={'var': -0.17486039103044})
-    X = transformer.fit_transform(dataframe_normal_dist)
-
-    df_transf = dataframe_normal_dist.copy()
-    df_transf['var'] = np.where(df_transf['var'] < -0.17486039103044, -0.17486039103044, df_transf['var'])
-
-    # init param
-    assert transformer.max_capping_dict is None
-    assert transformer.min_capping_dict == {'var': -0.17486039103044}
-    # fit params
-    assert transformer.right_tail_caps_ == {}
-    assert transformer.left_tail_caps_ == {'var': -0.17486039103044}
-    # transform params
-    pd.testing.assert_frame_equal(X, df_transf)
-    assert X['var'].min() >= -0.17486039103044
-    assert dataframe_normal_dist['var'].min() < -0.17486039103044
-
-    # test case 4: dataset contains na and transformer is asked to ignore them
-    transformer = ArbitraryOutlierCapper(max_capping_dict=None, min_capping_dict={'Age': 20},
-                                         missing_values='ignore')
-    X = transformer.fit_transform(dataframe_na)
-
-    df_transf = dataframe_na.copy()
-    df_transf['Age'] = np.where(df_transf['Age'] < 20, 20, df_transf['Age'])
-
-    # fit params
-    assert transformer.max_capping_dict is None
-    assert transformer.min_capping_dict == {'Age': 20}
-    assert transformer.input_shape_ == (8, 6)
-    # transform params
-    pd.testing.assert_frame_equal(X, df_transf)
-    assert X['Age'].min() >= 20
-    assert dataframe_na['Age'].min() < 20
-
-    with pytest.raises(ValueError):
-        ArbitraryOutlierCapper(max_capping_dict='other')
-
-    with pytest.raises(ValueError):
-        ArbitraryOutlierCapper(min_capping_dict='other')
-
-    with pytest.raises(ValueError):
-        ArbitraryOutlierCapper(min_capping_dict=None, max_capping_dict=None)
-
-    with pytest.raises(ValueError):
-        ArbitraryOutlierCapper(missing_values='other')
-
-    df_na = dataframe_normal_dist.copy()
-    df_na.loc[1, 'var'] = np.nan
-
-    # test case 5: when dataset contains na, fit method
-    with pytest.raises(ValueError):
-        transformer = ArbitraryOutlierCapper(min_capping_dict={'var': -0.17486039103044})
-        transformer.fit(df_na)
-
-    # test case 6: when dataset contains na, transform method
-    with pytest.raises(ValueError):
-        transformer = ArbitraryOutlierCapper(min_capping_dict={'var': -0.17486039103044})
-        transformer.fit(dataframe_normal_dist)
-        transformer.transform(df_na)
-
-    with pytest.raises(NotFittedError):
-        transformer = ArbitraryOutlierCapper(min_capping_dict={'var': -0.17486039103044})
-        transformer.transform(dataframe_vartypes)
-
-
-def test_OutlierTrimmer(dataframe_normal_dist, dataframe_na):
-    # test case 1: mean and std, right tail
-    transformer = OutlierTrimmer(distribution='gaussian', tail='right', fold=1)
-    X = transformer.fit_transform(dataframe_normal_dist)
-
-    df_transf = dataframe_normal_dist.copy()
-    outliers = np.where(df_transf['var'] > 0.10727677848029868, True, False)
-    df_transf = df_transf.loc[~outliers]
-
-    # transform params
-    pd.testing.assert_frame_equal(X, df_transf)
-    assert len(X) == 83
-
-    # test case 2: mean and std, both tails, different fold value
-    transformer = OutlierTrimmer(distribution='gaussian', tail='both', fold=2)
-    X = transformer.fit_transform(dataframe_normal_dist)
-    assert len(X) == 96
-
-    # test case 3: IQR, left tail, fold 2
-    transformer = OutlierTrimmer(distribution='skewed', tail='left', fold=0.8)
-    X = transformer.fit_transform(dataframe_normal_dist)
-
-    df_transf = dataframe_normal_dist.copy()
-    outliers = np.where(df_transf['var'] < -0.17486039103044, True, False)
-    df_transf = df_transf.loc[~outliers]
-
-    pd.testing.assert_frame_equal(X, df_transf)
-    assert len(X) == 98
-
-    # test case 4: dataset contains na, and transformer is asked to ignore
-    transformer = OutlierTrimmer(distribution='gaussian', tail='right', fold=1,
-                                 variables=['Age'], missing_values='ignore')
-    X = transformer.fit_transform(dataframe_na)
-
-    df_transf = dataframe_na.copy()
-    outliers = np.where(df_transf['Age'] > 38.79255087111844, True, False)
-    df_transf = df_transf.loc[~outliers]
-
-    pd.testing.assert_frame_equal(X, df_transf)
-    assert len(X) == 6
